1.Ders

+++ operatorler.py
@@ -15,22 +15,22 @@
 sayi1 = input("Lutfen 1. sayiyi giriniz : ")
 sayi2 = input("Lutfen 2. sayiyi giriniz : ")
 
-toplam = int(sayi1) + int(sayi2)
+toplam = float(sayi1) + float(sayi2)
 print("Toplama Islemi Sonucu : " + str(toplam))
 
-cikarma_islemi = int(sayi1) - int(sayi2)
+cikarma_islemi = float(sayi1) - float(sayi2)
 print("Cıkarma Islemi Sonucu : " + str(cikarma_islemi))
 
 bolme_islemi = float(sayi1) / float(sayi2)
 print("Bolme Islemi Sonucu : " + str(bolme_islemi))
 
-carpma_islemi = int(sayi1) * int(sayi2)
+carpma_islemi = float(sayi1) * float(sayi2)
 print("Carpma Islemi Sonucu : " + str(carpma_islemi))
 
 mod_islemi = float(sayi1) % float(sayi2)
 print("Mod Islemi Sonucu : " + str(mod_islemi))
 
-kuvvet_islemi = int(sayi1) ** int(sayi2)
+kuvvet_islemi = float(sayi1) ** float(sayi2)
 print("Kuvvet Islemi Sonucu : " + str(kuvvet_islemi))
